Move declarative Tool Contract tests out of root lib-test
