+++ crates/webcodex-tool-contracts/src/lib.rs
@@ -15,6 +15,9 @@ pub mod tool_spec;
 #[cfg(any(test, feature = "root-test-support"))]
 pub mod test_support;
 
+#[cfg(test)]
+mod tests;
+
 pub use metadata::*;
 pub use registry::*;
 pub use tool_catalog::*;
