+++ src/tool_runtime/tests/schema/definitions.rs
@@ -1,166 +1,6 @@
 use super::*;
 
 #[test]
-fn tool_definitions_cover_known_names_and_public_specs() {
-    use crate::tool_runtime::tool_definition::{
-        lookup_tool_definition, model_hidden_tool_names, model_visible_tool_definitions,
-        tool_definitions,
-    };
-
-    let definition_names = tool_definitions()
-        .map(|definition| definition.name)
-        .collect::<BTreeSet<_>>();
-    let definition_order = tool_definitions()
-        .map(|definition| definition.name)
-        .collect::<Vec<_>>();
-    let known_names = known_tool_names().collect::<BTreeSet<_>>();
-    let hidden_names = model_hidden_tool_names().collect::<BTreeSet<_>>();
-    let definition_hidden_names = tool_definitions()
-        .filter(|definition| definition.visibility.is_model_hidden())
-        .map(|definition| definition.name)
-        .collect::<BTreeSet<_>>();
-    for name in known_tool_names() {
-        assert!(
-            lookup_tool_definition(name).is_some(),
-            "{name} missing ToolDefinition lookup"
-        );
-    }
-    assert_eq!(
-        definition_names, known_names,
-        "ToolDefinition mirror must cover every ToolCall name exactly"
-    );
-    assert_eq!(
-        definition_order,
-        known_tool_names().collect::<Vec<_>>(),
-        "known-tool iterator must mirror canonical ToolDefinition order"
-    );
-    assert_eq!(
-        hidden_names, definition_hidden_names,
-        "hidden-name iterator must match ToolDefinition visibility"
-    );
-
-    let specs = registered_tool_specs();
-    let spec_names = specs
-        .iter()
-        .map(|spec| spec.name.as_str())
-        .collect::<BTreeSet<_>>();
-    let visible_definition_names = model_visible_tool_definitions()
-        .map(|definition| definition.name)
-        .collect::<BTreeSet<_>>();
-    let spec_order = specs
-        .iter()
-        .map(|spec| spec.name.clone())
-        .collect::<Vec<_>>();
-    let visible_definition_order = model_visible_tool_definitions()
-        .map(|definition| definition.name.to_string())
-        .collect::<Vec<_>>();
-    assert_eq!(
-        spec_names, visible_definition_names,
-        "model-visible ToolDefinitions must match public ToolSpecs"
-    );
-
-    assert_eq!(
-        visible_definition_order, spec_order,
-        "canonical ToolDefinition order must preserve public ToolSpec order"
-    );
-    assert_eq!(
-        registered_tool_names(),
-        visible_definition_order,
-        "public tool_names must derive from canonical model-visible ToolDefinition order"
-    );
-}
-
-#[test]
-fn adaptive_runtime_direct_declarations_are_visible_ranked_and_unique() {
-    use crate::tool_runtime::tool_definition::{
-        adaptive_runtime_direct_tool_definitions, tool_definitions,
-    };
-
-    let mut seen_ranks = std::collections::BTreeMap::new();
-    for definition in tool_definitions() {
-        let Some(rank) = definition.adaptive_runtime_direct_rank() else {
-            continue;
-        };
-        assert!(
-            definition.visibility.is_model_visible(),
-            "{} cannot declare adaptive direct rank {rank} while model-hidden",
-            definition.name
-        );
-        assert!(
-            seen_ranks.insert(rank, definition.name).is_none(),
-            "adaptive direct rank {rank} is duplicated"
-        );
-    }
-
-    let derived = adaptive_runtime_direct_tool_definitions();
-    assert!(
-        !derived.is_empty(),
-        "adaptive direct declarations must not be empty"
-    );
-    for pair in derived.windows(2) {
-        assert!(
-            pair[0].adaptive_runtime_direct_rank() < pair[1].adaptive_runtime_direct_rank(),
-            "adaptive direct definitions must be strictly ordered by unique rank"
-        );
-    }
-    assert_eq!(
-        derived.len(),
-        seen_ranks.len(),
-        "every adaptive direct declaration must be present in the derived set"
-    );
-    let apply_patch = derived
-        .iter()
-        .find(|definition| definition.name == "apply_patch")
-        .expect("apply_patch must be adaptive-direct");
-    let apply_text_edits = derived
-        .iter()
-        .find(|definition| definition.name == "apply_text_edits")
-        .expect("apply_text_edits must be adaptive-direct");
-    assert!(
-        apply_patch.adaptive_runtime_direct_rank()
-            < apply_text_edits.adaptive_runtime_direct_rank(),
-        "apply_patch must rank before apply_text_edits on the adaptive direct surface"
-    );
-}
-
-#[test]
-fn tool_definitions_drive_metadata_visibility_and_categories() {
-    use crate::tool_runtime::metadata::lookup_tool_metadata;
-    use crate::tool_runtime::tool_definition::tool_definitions;
-
-    for definition in tool_definitions() {
-        let metadata = definition.metadata();
-        let facade_metadata = lookup_tool_metadata(definition.name)
-            .copied()
-            .unwrap_or_else(|| panic!("{} missing metadata facade entry", definition.name));
-        assert_eq!(
-            metadata, facade_metadata,
-            "{} metadata facade must return ToolDefinition metadata",
-            definition.name
-        );
-        assert_eq!(metadata.name, definition.name);
-        assert_eq!(
-            definition.visibility.is_model_hidden(),
-            is_model_hidden_tool_name(definition.name),
-            "{} visibility mirror must match model-hidden filter",
-            definition.name
-        );
-        assert_eq!(
-            definition.category,
-            tool_manifest_category(definition.name),
-            "{} category mirror must match tool_manifest",
-            definition.name
-        );
-        assert_eq!(
-            definition.metadata().authority,
-            metadata.authority,
-            "{} authority mirror must match definition ToolMetadata",
-            definition.name
-        );
-    }
-}
-
-#[test]
 fn retired_delete_files_alias_stays_absent() {
     use crate::tool_runtime::metadata::lookup_tool_metadata;
     use crate::tool_runtime::tool_definition::lookup_tool_definition;
